Changed status code to be long, otherwise the call to curl_easy_getinfo() would overwrite the subsequent message pointer

+++ include/trial/http/curl/socket.hpp
@@ -133,7 +133,7 @@ private:
     struct
     {
         socket::state::value state;
-        int status_code;
+        long status_code;
         curl::message *message;
         curl::message storage;
     } current;
